localstoragedecuentasterminado

+++ 2perfil.js
@@ -2,7 +2,7 @@
 document.addEventListener("DOMContentLoaded", function () {
 
     // Toma la informacion de cuenta iniciada, la convierte de string a objeto
-    const storedCuenta =  JSON.parse(localStorage.getItem("cuentaIniciada"));
+    const storedCuenta = JSON.parse(localStorage.getItem("cuentaIniciada"));
     console.log(storedCuenta)
 
     // Lo guardado en el local storage como cuentaIniciada.usuario, se renderiza en username
@@ -10,13 +10,13 @@ document.addEventListener("DOMContentLoaded", function () {
         const usuarioButton = document.getElementById("username");
         usuarioButton.textContent = storedCuenta.usuario;
     }
-    
+
     // Lo guardado en el local storage como cuentaIniciada.email, se renderiza en email-placeholder
     if (storedCuenta) {
         const emailPlaceholder = document.getElementById("email-placeholder");
         emailPlaceholder.textContent = storedCuenta.email;
     }
-    
+
     // Lo guardado en el local storage como cuentaIniciada.telefono, se renderiza en telefono-placeholder
     if (storedCuenta) {
         const telefonoPlaceholder = document.getElementById("telefono-placeholder");
@@ -60,28 +60,36 @@ const modal4 = document.getElementById("myModal4");
 const guardarCambiosBtn4 = document.getElementById("guardar-cambios4");
 const usuarioInput = document.getElementById("input4");
 
-//Funcion para guardar el nuevo email proporcionado en el input del modal por el usuario
+// Funcion para guardar el nuevo email proporcionado en el input del modal por el usuario
 guardarCambiosBtn.addEventListener("click", () => {
     const newemail = emailInput.value;
-    //El email tiene que tener texto de cualquier tipo+@ con otro texto + . y texto
+    // El email tiene que tener texto de cualquier tipo + @ con otro texto + . y texto
     const emailPattern = /^[a-zA-Z0-9._-]+@[a-zA-Z0-9.-]+\.[a-zA-Z]{2,4}$/;
 
-    // SI la condiccion no se cumple aparece una alerta que advierte al usuario
+    // Si la condición no se cumple aparece una alerta que advierte al usuario
     if (!emailPattern.test(newemail)) {
         alert("Por favor, ingresa una dirección de correo electrónico válida.");
         return;
     }
-   
-    var storedCuenta =  JSON.parse(localStorage.getItem("cuentaIniciada"));
+
+    // Verificar que el nuevo correo electrónico sea único
     var cuentas = JSON.parse(localStorage.getItem("cuentas"));
+    const emailExistente = cuentas.some(account => account.email === newemail);
+
+    if (emailExistente) {
+        alert("Este correo electrónico ya está en uso. Por favor, elige otro.");
+        return;
+    }
+
+    // Continuar con la actualización si el nuevo correo es único
+    var storedCuenta = JSON.parse(localStorage.getItem("cuentaIniciada"));
     const cuentaPorEmail = cuentas.find(account => account.email === storedCuenta.email);
-    cuentaPorEmail.email=newemail;
+    cuentaPorEmail.email = newemail;
     localStorage.setItem("cuentas", JSON.stringify(cuentas));
     storedCuenta.email = newemail;
     localStorage.setItem("cuentaIniciada", JSON.stringify(storedCuenta));
 
-                
-    //Lo recien guardado en el local storage como email, se renderiza en email-placeholder
+    // Lo recién guardado en el local storage como email, se renderiza en email-placeholder
     const emailPlaceholder = document.getElementById("email-placeholder");
     emailPlaceholder.textContent = storedCuenta.email;
 });
@@ -96,11 +104,11 @@ guardarCambiosBtn2.addEventListener("click", () => {
         alert("El número de teléfono debe contener exactamente 10 números.");
         return;
     }
-    
-    var storedCuenta =  JSON.parse(localStorage.getItem("cuentaIniciada"));
+
+    var storedCuenta = JSON.parse(localStorage.getItem("cuentaIniciada"));
     var cuentas = JSON.parse(localStorage.getItem("cuentas"));
     const cuentaPorTelefono = cuentas.find(account => account.telefono === storedCuenta.telefono);
-    cuentaPorTelefono.telefono=newtelefono;
+    cuentaPorTelefono.telefono = newtelefono;
     localStorage.setItem("cuentas", JSON.stringify(cuentas));
     storedCuenta.telefono = newtelefono;
     localStorage.setItem("cuentaIniciada", JSON.stringify(storedCuenta));
@@ -119,15 +127,15 @@ guardarCambiosBtn3.addEventListener("click", () => {
     const contrasenaCapital = /[A-Z]/.test(newcontrasena);
     //la contrasena debe tener minimo 8 caracteres
     // SI las condiccions no se cumple aparece una alerta que advierte al usuario y lo pone al tanto de la estructura requerida
-    if (newcontrasena.length <= 8 || !contrasenaPattern || !contrasenaCapital) {
+    if (newcontrasena.length <= 7 || !contrasenaPattern || !contrasenaCapital) {
         alert("La contraseña debe tener al menos 8 caracteres con al menos un número y al menos una letra mayúscula.");
         return;
     }
 
-    var storedCuenta =  JSON.parse(localStorage.getItem("cuentaIniciada"));
+    var storedCuenta = JSON.parse(localStorage.getItem("cuentaIniciada"));
     var cuentas = JSON.parse(localStorage.getItem("cuentas"));
     const cuentaPorContrasena = cuentas.find(account => account.contrasena === storedCuenta.contrasena);
-    cuentaPorContrasena.contrasena=newcontrasena;
+    cuentaPorContrasena.contrasena = newcontrasena;
     localStorage.setItem("cuentas", JSON.stringify(cuentas));
     storedCuenta.contrasena = newcontrasena;
     localStorage.setItem("cuentaIniciada", JSON.stringify(storedCuenta));
@@ -137,25 +145,34 @@ guardarCambiosBtn3.addEventListener("click", () => {
     contrasenaPlaceholder.textContent = storedCuenta.contrasena;
 });
 
-//Funcion para guardar el nuevo usuario  proporcionado en el input del modal por el usuario
+// Funcion para guardar el nuevo usuario proporcionado en el input del modal por el usuario
 guardarCambiosBtn4.addEventListener("click", () => {
     const newusuario = usuarioInput.value;
-    //El usuario tienen que tener entre 6 y 12 caracteres.
-    // SI la condiccion no se cumple aparece una alerta que advierte al usuario y lo pone al tanto de la estructura requerida
+    // El usuario debe tener entre 6 y 12 caracteres.
+    // Si la condición no se cumple aparece una alerta que advierte al usuario y lo pone al tanto de la estructura requerida
     if (newusuario.length < 6 || newusuario.length > 12) {
         alert("El usuario debe tener entre 6 y 12 caracteres.");
         return;
     }
 
-    var storedCuenta =  JSON.parse(localStorage.getItem("cuentaIniciada"));
+    // Verificar que el nuevo nombre de usuario sea único
     var cuentas = JSON.parse(localStorage.getItem("cuentas"));
+    const usuarioExistente = cuentas.some(account => account.usuario === newusuario);
+
+    if (usuarioExistente) {
+        alert("Este nombre de usuario ya está en uso. Por favor, elige otro.");
+        return;
+    }
+
+    // Continuar con la actualización si el nuevo nombre de usuario es único
+    var storedCuenta = JSON.parse(localStorage.getItem("cuentaIniciada"));
     const cuentaPorUsuario = cuentas.find(account => account.usuario === storedCuenta.usuario);
-    cuentaPorUsuario.usuario=newusuario;
+    cuentaPorUsuario.usuario = newusuario;
     localStorage.setItem("cuentas", JSON.stringify(cuentas));
     storedCuenta.usuario = newusuario;
     localStorage.setItem("cuentaIniciada", JSON.stringify(storedCuenta));
 
-    //Lo recien guardado en el local storage como usuario, se renderiza en username
+    // Lo recién guardado en el local storage como usuario, se renderiza en username
     const usuarioButton = document.getElementById("username");
     usuarioButton.textContent = storedCuenta.usuario;
 });
@@ -233,11 +250,11 @@ window.addEventListener("click", (event) => {
 });
 
 // Función que se ejecuta al oprimir el botón de Cerrar sesion
-function cerrarSesion(){
+function cerrarSesion() {
 
-            //Borrar la informacion de cuenta iniciada del local storage
-            localStorage.removeItem("cuentaIniciada");
-            
-            // Redirige al usuario a la página de bienvenida
-            window.location.href = "./index.html";
+    //Borrar la informacion de cuenta iniciada del local storage
+    localStorage.removeItem("cuentaIniciada");
+
+    // Redirige al usuario a la página de bienvenida
+    window.location.href = "./index.html";
 };
